remove most next direct deps
