+++ .changeset/config.json
@@ -7,5 +7,5 @@
   "access": "public",
   "baseBranch": "origin/main",
   "updateInternalDependencies": "patch",
-  "ignore": ["example", "docs", "database"]
+  "ignore": ["example", "docs", "database", "examples-common"]
 }
