+++ apps/example-start/app/routes/admin.$.tsx
@@ -1,6 +1,7 @@
 import { AdminComponentProps, NextAdmin } from "@premieroctet/next-admin";
 import { NextAdminRouterAdapter } from "@premieroctet/next-admin/adapters/tanstack-router";
 import { createFileRoute } from "@tanstack/react-router";
+import { Dashboard } from "examples-common/components";
 import { getNextAdminPropsFn } from "../functions/nextadmin";
 import { options } from "../options";
 
@@ -18,13 +19,26 @@ export const Route = createFileRoute("/admin/$")({
 function RouteComponent() {
   const data = Route.useLoaderData();
 
+  const logoutRequest: [RequestInfo, RequestInit] = [
+    "/",
+    {
+      method: "POST",
+    },
+  ];
+
   return (
     <NextAdminRouterAdapter>
       <NextAdmin
         // @ts-expect-error
         {...(data.props as AdminComponentProps)}
-        dashboard={<div></div>}
+        dashboard={<Dashboard />}
         options={options}
+        user={{
+          data: {
+            name: "John Doe",
+          },
+          logout: logoutRequest,
+        }}
       />
     </NextAdminRouterAdapter>
   );
