+++ apps/example/app/[locale]/admin/[[...nextadmin]]/page.tsx
@@ -1,11 +1,13 @@
-import Dashboard from "@/components/Dashboard";
+import { Dashboard } from "examples-common/components";
 import { options } from "@/options";
 import { prisma } from "@/prisma";
 import { NextAdmin, PromisePageProps } from "@premieroctet/next-admin";
 import { NextAdminRouterAdapter } from "@premieroctet/next-admin/adapters/next";
 import { getNextAdminProps } from "@premieroctet/next-admin/appRouter";
+import PageLoader from "@premieroctet/next-admin/pageLoader";
 import { Metadata, Viewport } from "next";
 import { getMessages } from "next-intl/server";
+import "react-datepicker/dist/react-datepicker.css";
 
 export const viewport: Viewport = {
   initialScale: 1,
@@ -52,6 +54,7 @@ export default async function AdminPage(props: PromisePageProps) {
           },
           logout: logoutRequest,
         }}
+        pageLoader={<PageLoader />}
       />
     </NextAdminRouterAdapter>
   );
