+++ apps/example/next.config.js
@@ -14,6 +14,6 @@ module.exports = withNextIntl(
           },
         ],
       ],
-    }
+    },
   })
 );
