+++ apps/example/package.json
@@ -20,6 +20,7 @@
     "@premieroctet/next-admin": "workspace:*",
     "@prisma/client": "5.14.0",
     "@tremor/react": "^3.2.2",
+    "examples-common": "workspace:*",
     "next": "^15.1.0",
     "next-intl": "^3.3.2",
     "next-superjson": "^1.0.7",
