+++ apps/example/pages/pagerouter/admin/[[...nextadmin]].tsx
@@ -1,4 +1,6 @@
 import { AdminComponentProps, NextAdmin } from "@premieroctet/next-admin";
+import { NextAdminRouterAdapter } from "@premieroctet/next-admin/adapters/next";
+import PageLoader from "@premieroctet/next-admin/pageLoader";
 import { getNextAdminProps } from "@premieroctet/next-admin/pageRouter";
 import { GetServerSideProps } from "next";
 import { options } from "../../../pageRouterOptions";
@@ -9,16 +11,19 @@ const pageOptions = options;
 
 export default function Admin(props: AdminComponentProps) {
   return (
-    <NextAdmin
-      {...props}
-      options={pageOptions}
-      user={{
-        data: {
-          name: "John Doe",
-        },
-        logout: ["/"],
-      }}
-    />
+    <NextAdminRouterAdapter>
+      <NextAdmin
+        {...props}
+        options={pageOptions}
+        user={{
+          data: {
+            name: "John Doe",
+          },
+          logout: ["/"],
+        }}
+        pageLoader={<PageLoader />}
+      />
+    </NextAdminRouterAdapter>
   );
 }
