+++ packages/examples-common/actions/addTag.js
@@ -1,0 +1,17 @@
+"use server";
+import prisma from "database";
+const addTag = async (tag, selectedIds) => {
+    await prisma.post.updateMany({
+        where: {
+            id: {
+                in: selectedIds,
+            },
+        },
+        data: {
+            tags: {
+                push: tag,
+            },
+        },
+    });
+};
+export default addTag;
