+++ packages/examples-common/components/InputDisplay.tsx
