+++ packages/examples-common/components/PasswordInput.tsx
