+++ packages/examples-common/components/index.ts
@@ -1,0 +1,6 @@
+export { default as Dashboard } from "./Dashboard";
+export { default as DatePicker } from "./DatePicker";
+export { default as InputDisplay } from "./InputDisplay";
+export { default as PasswordInput } from "./PasswordInput";
+export { default as PostAddTagDialogContent } from "./PostAddTagDialogContent";
+export { default as UserDetailsDialogContent } from "./UserDetailsDialogContent";
